+++ src/Vankosoft/ApplicationBundle/Component/Context/RequestResolverInterface.php
@@ -1,9 +1,8 @@
 <?php namespace Vankosoft\ApplicationBundle\Component\Context;
 
-use Symfony\Component\HttpFoundation\Request;
 use Vankosoft\ApplicationBundle\Model\Interfaces\ApplicationInterface;
 
 interface RequestResolverInterface
 {
-    public function findApplication( Request $request ) : ?ApplicationInterface;
+    public function findApplication( string $host ) : ?ApplicationInterface;
 }